refactor:  상세보기 닫기 버튼 수정
