+++ src/pages/lanternDetail/LanternDetail.css.ts
@@ -16,21 +16,6 @@ export const fullImage = style({
   objectFit: 'cover',
 });
 
-export const closeButton = style({
-  position: 'absolute',
-  top: 20,
-  left: 20,
-  zIndex: 1000,
-  width: 100,
-  height: 100,
-  backgroundColor: 'rgba(255, 255, 255, 0.7)',
-  border: 'none',
-  borderRadius: 4,
-  padding: '8px 12px',
-  fontSize: 16,
-  cursor: 'pointer',
-});
-
 export const handPointer = style({
   position: 'fixed',
   width: 20,
